+++ common/src/main/java/dev/boxadactle/boxlib/gui/config/widget/button/BColorPickerButton.java
@@ -15,6 +15,7 @@ import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.layouts.*;
 import net.minecraft.client.gui.narration.NarrationElementOutput;
 import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.input.MouseButtonEvent;
 import net.minecraft.network.chat.Component;
 
 import java.util.function.Consumer;
@@ -58,7 +59,7 @@ public class BColorPickerButton extends BOptionButton<Integer> {
     }
 
     @Override
-    public void onClick(double mouseX, double mouseY) {
+    public void onClick(MouseButtonEvent mouseButtonEvent, boolean bl) {
         ClientUtils.setScreen(new ColorPickerScreen(getMessage(), currentValue));
     }